delete admin

+++ backend/api/admins.js
@@ -33,11 +33,18 @@ router.post("/user", async (req, res) => {
 
     const socket = req.app.get("socket");
 
+    const isAdmin = await adminUserModel.findOne({
+      admin: user.cmuAccount,
+    })
+    if(!isAdmin){
+      return res.send({ok: false, message: "Your account isn't admin!" })
+    }
+
     const admin = await adminUserModel.findOne({
       admin: req.body.admin,
     });
 
-    if (admin) return res.send("Admin user already exist.");
+    if (admin) return res.send({ok: true, message: "Admin user already exist"});
 
     await adminUserModel.create({
       admin: req.body.admin,
@@ -63,6 +70,13 @@ router.delete("/user", async (req, res) => {
 
     const socket = req.app.get("socket");
 
+    const isAdmin = await adminUserModel.findOne({
+      admin: user.cmuAccount,
+    })
+    if(!isAdmin){
+      return res.send({ok: false, message: "Your account isn't admin!" })
+    }
+
     await adminUserModel.findOneAndDelete({
       _id: req.query._id,
     });
@@ -78,9 +92,21 @@ router.delete("/user", async (req, res) => {
 
 router.post("/delete", async (req, res) => {
   try {
-    const socket = req.app.get("socket");
-    const _id = req.body._id;
+    const token = req.cookies.token;
+    const user = await verifyAndValidateToken(token, res);
+    if (!user.cmuAccount)
+      return res.status(403).send({ ok: false, message: "Invalid token" });
 
+    const socket = req.app.get("socket");
+
+    const isAdmin = await adminUserModel.findOne({
+      admin: user.cmuAccount,
+    })
+    if(!isAdmin){
+      return res.send({ok: false, message: "Your account isn't admin!" })
+    }
+
+    const _id = req.body._id;
     const current = await adminModel.findOneAndDelete({ _id: _id });
 
     socket.emit("currentUpdate", current);
@@ -101,6 +127,14 @@ router.post("/", async (req, res) => {
       return res.status(403).send({ ok: false, message: "Invalid token" });
 
     const socket = req.app.get("socket");
+
+    const isAdmin = await adminUserModel.findOne({
+      admin: user.cmuAccount,
+    })
+    if(!isAdmin){
+      return res.send({ok: false, message: "Your account isn't admin!" })
+    }
+    
     const semester = parseInt(req.body.semester);
     const year = parseInt(req.body.year);
